+++ tests/RdKafka/MessageTest.php
@@ -27,7 +27,7 @@ class MessageTest extends TestCase
         $consumerTopic = $consumer->newTopic(KAFKA_TEST_TOPIC);
         $consumerTopic->consumeStart(0, rd_kafka_offset_tail(1));
 
-        $this->message = $consumerTopic->consume(0, KAFKA_TEST_TIMEOUT_MS);
+        $this->message = $consumerTopic->consume(0, (int) KAFKA_TEST_TIMEOUT_MS);
 
         $consumerTopic->consumeStop(0);
     }
